Creation des services et ajout des exceptions

+++ src/main/java/com/groupeisi/commerceenligne/dao/IAchatRepository.java
@@ -3,5 +3,5 @@ package com.groupeisi.commerceenligne.dao;
 import com.groupeisi.commerceenligne.entities.Achats;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAchatRepository extends JpaRepository<Achats,Integer> {
+public interface IAchatRepository extends JpaRepository<Achats,Long> {
 }
