+++ src/main/java/com/groupeisi/commerceenligne/dao/IVenteRepository.java
@@ -3,5 +3,5 @@ package com.groupeisi.commerceenligne.dao;
 import com.groupeisi.commerceenligne.entities.Ventes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IVenteRepository extends JpaRepository<Ventes,Integer> {
+public interface IVenteRepository extends JpaRepository<Ventes,Long> {
 }
